requirements-test.txt
pytest
pytest-mock
requests-mock
python-dotenv
